requirements.txt
numpy
Pillow
PyInstaller
pyscreenshot
pywin32